requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tabulate>=0.9.0
yfinance>=0.2.30
pytest>=7.4.0
